Add tabview overlay actions
