+++ FlipperZero/UI/Sources/View/Common/ShareView.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+func share(_ items: [Any]) {
+    let activityContoller = UIActivityViewController(
+        activityItems: items,
+        applicationActivities: nil)
+    UIApplication.shared
+        .windows
+        .first?
+        .rootViewController?
+        .present(activityContoller, animated: true, completion: nil)
+}
